Conflicts:
	.DEPS.git

+++ webkit/dom_storage/dom_storage_context.cc
@@ -46,6 +46,18 @@ DomStorageContext::DomStorageContext(
 }
 
 DomStorageContext::~DomStorageContext() {
+  if (session_storage_database_.get()) {
+    // SessionStorageDatabase shouldn't be deleted right away: deleting it will
+    // potentially involve waiting in leveldb::DBImpl::~DBImpl, and waiting
+    // shouldn't happen on this thread. This will unassign the ref counted
+    // pointer without decreasing the ref count, and is balanced by the Release
+    // that happens later.
+    task_runner_->PostShutdownBlockingTask(
+        FROM_HERE,
+        DomStorageTaskRunner::COMMIT_SEQUENCE,
+        base::Bind(&SessionStorageDatabase::Release,
+                   base::Unretained(session_storage_database_.release())));
+  }
 }
 
 DomStorageNamespace* DomStorageContext::GetStorageNamespace(
@@ -292,7 +304,8 @@ void DomStorageContext::StartScavengingUnusedSessionStorage() {
 
 void DomStorageContext::FindUnusedNamespaces() {
   DCHECK(session_storage_database_.get());
-  DCHECK(!scavenging_started_);
+  if (scavenging_started_)
+    return;
   scavenging_started_ = true;
   std::set<std::string> namespace_ids_in_use;
   for (StorageNamespaceMap::const_iterator it = namespaces_.begin();
@@ -343,6 +356,8 @@ void DomStorageContext::DeleteNextUnusedNamespace() {
 }
 
 void DomStorageContext::DeleteNextUnusedNamespaceInCommitSequence() {
+  if (deletable_persistent_namespace_ids_.empty())
+    return;
   const std::string& persistent_id = deletable_persistent_namespace_ids_.back();
   session_storage_database_->DeleteNamespace(persistent_id);
   deletable_persistent_namespace_ids_.pop_back();
